backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
